Day 2: Diffuse reflections.
For some reason this image is much darker than the corresponding image
from the book. The sky color is a perfect match so the only explanation
I can see is that the distribution of scattered rays is different
leading to more rays contributing zero to the final color of the
spheres.